Add files via upload
I think this chording is a fairly good interpretation on the melody. The original guitarwork may be just power chords, they don't bother with thirds at all, but the melody makes the minor chords quite obvious to my ear, at least. I think the bass line progression in "I think I'm cracking up" is D-C#-B (and I'd instinctively put an A major under the C#). I find it funny that most of the chords I googled up seem to either get it wrong or ignore it completely.